# backend/swagger.yaml
openapi: 3.0.3
info:
  title: VRMS API
  description: |
    Volunteer Relationship Management System (VRMS) API for Hack for LA.

    ## Authentication
    Most endpoints require the custom header `x-customrequired-header`.
    Auth endpoints use JWT tokens via `x-access-token` header or `Authorization: Bearer <token>`,
    and refresh tokens stored in httpOnly cookies.

    **Exempt from custom header:**
    - `GET /api/recurringevents`
    - `GET /api/healthcheck`
  version: 1.0.0
  contact:
    name: Hack for LA
    url: https://www.hackforla.org
servers:
  - url: /api
    description: API base path
...
components:
  securitySchemes:
    CustomHeader:
      type: apiKey
      in: header
      name: x-customrequired-header
      description: Required custom header for most endpoints
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token
    AccessTokenHeader:
      type: apiKey
      in: header
      name: x-access-token
      description: JWT access token (alternative to Bearer)
    CookieAuth:
      type: apiKey
      in: cookie
      name: token
      description: JWT token stored in httpOnly cookie
    RefreshTokenCookie:
      type: apiKey
      in: cookie
      name: refresh_token
      description: Refresh token stored in httpOnly cookie
  schemas:
...
    Error:
      type: object
      properties:
        message:
          type: string
security:
  - CustomHeader: []
paths:
...
  /grantpermission:
    post:
      tags: [GrantPermission]
      summary: OAuth flow for Google Drive permission
      description: |
        Handles the OAuth2 flow for granting Google Drive permissions.
        - If `token` and `code` are provided, completes the OAuth flow.
        - Otherwise, returns an authorization URL to start the flow.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, file]
              properties:
                token:
                  type: string
                  description: Existing OAuth token (optional)
                code:
                  type: string
                  description: OAuth authorization code (optional)
                email:
                  type: string
                  format: email
                file:
                  type: string
                  description: Google Drive file ID
      responses:
        '200':
          description: Permission granted or auth URL returned
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    properties:
                      message:
                        type: string
                        example: Success!
                      token:
                        type: string
                  - type: object
                    properties:
                      url:
                        type: string
                        format: uri
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /grantpermission/googleDrive:
    post:
      tags: [GrantPermission]
      summary: Grant Google Drive permission directly
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, file]
              properties:
                email:
                  type: string
                  format: email
                file:
                  type: string
                  description: Google Drive file ID
      responses:
        '200':
          description: Permission granted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Success!
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /grantpermission/gitHub:
    post:
      tags: [GrantPermission]
      summary: Onboard user to GitHub org and team
      description: |
        Invites user to the GitHub organization and adds them to a team.
        Requires JWT authentication.
      security:
        - CustomHeader: []
          BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [teamName, handle]
              properties:
                teamName:
                  type: string
                  description: GitHub team slug
                handle:
                  type: string
                  description: GitHub username
      responses:
        '200':
          description: GitHub onboarding result
          content:
            application/json:
              schema:
                type: object
                properties:
                  orgMembershipStatus:
                    type: string
                    description: Organization membership status
                  teamMembershipStatus:
                    type: string
                    description: Team membership status
                  publicMembership:
                    type: string
                    description: Public membership visibility (if applicable)
                  twoFAenabled:
                    type: boolean
                    description: Whether 2FA is enabled (if applicable)
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ─── Health Check ──────────────────────────────────────
...